⚡prettier formatting

+++ library/AddArrayLib.sol
@@ -5,7 +5,7 @@ library AddrArrayLib {
     using AddrArrayLib for Addresses;
 
     struct Addresses {
-      address[]  _items;
+        address[] _items;
     }
 
     /**
@@ -15,9 +15,9 @@ library AddrArrayLib {
      * @param element the element to add in the array
      */
     function pushAddress(Addresses storage self, address element) internal {
-      if (!exists(self, element)) {
-        self._items.push(element);
-      }
+        if (!exists(self, element)) {
+            self._items.push(element);
+        }
     }
 
     /**
@@ -28,7 +28,7 @@ library AddrArrayLib {
      * @param element the element to remove from the array
      */
     function removeAddress(Addresses storage self, address element) internal {
-        for (uint i = 0; i < self.size(); i++) {
+        for (uint256 i = 0; i < self.size(); i++) {
             if (self._items[i] == element) {
                 self._items[i] = self._items[self.size() - 1];
                 self._items.pop();
@@ -42,7 +42,11 @@ library AddrArrayLib {
      * @param self Storage array containing address type variables
      * @param index the index in the array
      */
-    function getAddressAtIndex(Addresses memory self, uint256 index) internal view returns (address) {
+    function getAddressAtIndex(Addresses memory self, uint256 index)
+        internal
+        view
+        returns (address)
+    {
         require(index < size(self), "INVALID_INDEX");
         return self._items[index];
     }
@@ -52,7 +56,7 @@ library AddrArrayLib {
      * @param self Storage array containing address type variables
      */
     function size(Addresses memory self) internal view returns (uint256) {
-      return self._items.length;
+        return self._items.length;
     }
 
     /**
@@ -60,8 +64,12 @@ library AddrArrayLib {
      * @param self Storage array containing address type variables
      * @param element the element to check if it exists in the array
      */
-    function exists(Addresses memory self, address element) internal view returns (bool) {
-        for (uint i = 0; i < self.size(); i++) {
+    function exists(Addresses memory self, address element)
+        internal
+        view
+        returns (bool)
+    {
+        for (uint256 i = 0; i < self.size(); i++) {
             if (self._items[i] == element) {
                 return true;
             }
@@ -73,8 +81,11 @@ library AddrArrayLib {
      * @notice get the array
      * @param self Storage array containing address type variables
      */
-    function getAllAddresses(Addresses memory self) internal view returns(address[] memory) {
+    function getAllAddresses(Addresses memory self)
+        internal
+        view
+        returns (address[] memory)
+    {
         return self._items;
     }
-
 }
